requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
firecrawl-py
jinja2
aiofiles
